Rename codec to encode
