+++ ethereum_history_api/oracles/src/noir/oracles/accountOracle/encode.ts
